+++ src/pages/Upsolve/CarouselUpsolve.js
@@ -14,7 +14,11 @@ export default function Carousel_Upsolve(props) {
           <CommonCard name={props.name} />
         </Col>
         <Col span={19}>
-          <Carousel responsive={RESPONSIVE}>
+          
+          <Carousel
+            customTransition="all .5"
+            transitionDuration={500}
+            responsive={RESPONSIVE}>
             {props.problems.map((prob) => {
               if (prob.status === 'solved') {
                 if (props.notAttemptedToggle == false) {
@@ -22,7 +26,7 @@ export default function Carousel_Upsolve(props) {
                     <Col span={19}>
                       <div className="solved">
                         <CommonQues
-                          platform={props.platform}
+                          sitename={props.sitename}
                           url={prob.url}
                           index={prob.index}
                           name={prob.name}
@@ -34,30 +38,34 @@ export default function Carousel_Upsolve(props) {
                     </Col>
                   )
                 }
+                
               } else if (prob.status === 'wrong') {
-                return (
-                  <Col span={19}>
-                    {' '}
-                    <div className="wrong">
-                      <CommonQues
-                        platform={props.platform}
-                        url={prob.url}
-                        index={prob.index}
-                        name={prob.name}
-                        className="red"
-                        tags={prob.tags}
-                        status="WRONG"
-                      />
-                    </div>
-                  </Col>
-                )
+                
+                  return (
+                    <Col span={19}>
+                      {' '}
+                      <div className="wrong">
+                        <CommonQues
+                          sitename={props.sitename}
+                          url={prob.url}
+                          index={prob.index}
+                          name={prob.name}
+                          className="red"
+                          tags={prob.tags}
+                          status="WRONG"
+                        />
+                      </div>
+                    </Col>
+                  )
+                
+
               } else if (prob.status === 'upsolved') {
                 if (props.notAttemptedToggle == false) {
                   return (
                     <Col span={19}>
                       <div className="upsolved">
                         <CommonQues
-                          platform={props.platform}
+                          sitename={props.sitename}
                           url={prob.url}
                           index={prob.index}
                           name={prob.name}
@@ -75,7 +83,7 @@ export default function Carousel_Upsolve(props) {
                     {' '}
                     <div className="not_attempted">
                       <CommonQues
-                        platform={props.platform}
+                        sitename={props.sitename}
                         url={prob.url}
                         index={prob.index}
                         name={prob.name}
